backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.15.1
gunicorn==20.1.0
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2023.3
whitenoise==6.4.0
